fix: corrigir todos os jobs Bronze→Silver e Silver→Gold para execução real na AWS
Bronze → Silver (18 jobs):
- F.col("Op") → F.col("op"): Glue Catalog normaliza nomes de coluna para minúsculo
- Adicionado SILVER_COLS + df_valid.select(*SILVER_COLS) antes de cada MERGE INTO
  para evitar inserção de colunas extras do Bronze na tabela Silver
- job_tb_avaliacao_qualidade: cast explícito de id_supervisor para LongType()
  antes do coalesce para evitar inferência StringType → falha de escrita BIGINT
- job_tb_cliente: formato dt_cadastro corrigido para yyyy-MM-dd (era yyyy-MM-dd HH:mm:ss)
- job_tb_metricas_operacionais: nr_nivel_servico calculado (não existe no Bronze);
  dt_referencia formato yyyy-MM-dd; drop de id_operador e nr_nps
- job_tb_fila_atendimento: nr_sla_segundos/minutos criados como NULL (não existem no Bronze)
- job_tb_discagem: nr_telefone não existe no Bronze → nr_telefone_mascarado = ''

Silver → Gold (22 jobs):
- dim_cliente/operador/campanha/data: substituído spark.createDataFrame() com None
  em colunas de data por spark.sql() com tipos explícitos (evita ValueError de inferência)
- job_fato_chamada: removido JOIN inválido tp_chamada == nm_canal (ENTRADA/SAÍDA ≠
  TELEFONE/CHAT); substituído por lookup pontual WHERE nm_canal='TELEFONE'
- job_fato_qualidade: sk_chamada INT → BIGINT (overflow: monotonically_increasing_id
  do fato_chamada produz valores BIGINT que não cabem em INT)
- job_fato_interacao_ticket: sk_ticket INT → BIGINT (mesmo motivo, herdado de fato_ticket)
- job_fato_mensagem_chat: sk_chat INT → BIGINT (mesmo motivo, herdado de fato_chat)

Documentação:
- README.md: seção de validação com contagens reais do Athena (dims + fatos),
  integridade referencial (0 orphans), distribuição de qualidade; flags do
  06_run_pipeline.py atualizados (--only dims/fatos)
- pipeline/bronze_to_silver/README.md: correção das faixas de nota, colunas
  ausentes no Bronze, notas de implementação (CDC lowercase, bookmark vazio)
- pipeline/silver_to_gold/README.md: padrão de SK corrigido (dims=row_number INT,
  fatos=monotonically_increasing_id BIGINT); seção de bugs corrigidos em execução

+++ pipeline/silver_to_gold/job_fato_interacao_ticket_gold.py
@@ -73,7 +73,7 @@ df_fato = (
           F.to_date(df_interacao["dt_interacao"]) == F.col("_dt"), how="left")
 
     .withColumn("sk_ticket",
-        F.coalesce(F.col("_sk_ticket"), F.lit(-1).cast("int")))
+        F.coalesce(F.col("_sk_ticket"), F.lit(-1).cast("long")))
     .withColumn("sk_operador",
         F.coalesce(F.col("_sk_op"),     F.lit(-1).cast("int")))
     .withColumn("sk_data",
@@ -103,7 +103,7 @@ spark.sql(f"""
     CREATE TABLE IF NOT EXISTS glue_catalog.{GOLD_TABLE} (
         sk_interacao                    BIGINT,
         nk_interacao                    BIGINT,
-        sk_ticket                       INT,
+        sk_ticket                       BIGINT,
         sk_operador                     INT,
         sk_data                         INT,
         ds_canal                        STRING,
